+++ src/sll/init.c
@@ -7,6 +7,9 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/vm.h>
+#ifdef DEBUG_BUILD
+#include <signal.h>
+#endif
 
 
 
@@ -79,6 +82,9 @@ __SLL_NO_RETURN void _force_exit(const sll_char_t* a,const sll_char_t* b,const s
 			sll_platform_file_write(fd,">\n",2,NULL);
 		}
 	}
+#ifdef DEBUG_BUILD
+	raise(SIGABRT);
+#endif
 	_force_exit_platform();
 }